# Optimize partition cache getRegionReplicaSet interface performance by batching