--- a/iotdb-core/datanode/src/main/java/org/apache/iotdb/db/queryengine/plan/analyze/cache/partition/PartitionCache.java
+++ b/iotdb-core/datanode/src/main/java/org/apache/iotdb/db/queryengine/plan/analyze/cache/partition/PartitionCache.java
@@ -72,7 +72,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -543,6 +543,9 @@ public class PartitionCache {
    * @throws StatementAnalyzeException if there are exception when try to get latestRegionRouteMap
    */
   public List<TRegionReplicaSet> getRegionReplicaSet(List<TConsensusGroupId> consensusGroupIds) {
+    if (consensusGroupIds.isEmpty()) {
+      return Collections.emptyList();
+    }
     List<TRegionReplicaSet> result;
     // try to get regionReplicaSet from cache
     regionReplicaSetLock.readLock().lock();
@@ -730,14 +733,17 @@ public class PartitionCache {
           new HashMap<>();
       Map<TSeriesPartitionSlot, TRegionReplicaSet> regionReplicaSetMap =
           schemaPartitionMap.computeIfAbsent(database, k -> new HashMap<>());
-      List<TRegionReplicaSet> regionReplicaSets =
-          getRegionReplicaSet(
-              new ArrayList<>(schemaPartitionTable.getSchemaPartitionMap().values()));
-      Iterator<TRegionReplicaSet> iterator = regionReplicaSets.iterator();
-      for (Map.Entry<TSeriesPartitionSlot, TConsensusGroupId> entry :
-          schemaPartitionTable.getSchemaPartitionMap().entrySet()) {
-        regionReplicaSetMap.put(entry.getKey(), iterator.next());
+
+      Map<TSeriesPartitionSlot, TConsensusGroupId> orderedMap =
+          new LinkedHashMap<>(schemaPartitionTable.getSchemaPartitionMap());
+      List<TConsensusGroupId> orderedGroupIds = new ArrayList<>(orderedMap.values());
+      List<TRegionReplicaSet> regionReplicaSets = getRegionReplicaSet(orderedGroupIds);
+
+      int index = 0;
+      for (Map.Entry<TSeriesPartitionSlot, TConsensusGroupId> entry : orderedMap.entrySet()) {
+        regionReplicaSetMap.put(entry.getKey(), regionReplicaSets.get(index++));
       }
+
       logger.debug("[{} Cache] hit", CacheMetrics.SCHEMA_PARTITION_CACHE_NAME);
       // cache hit
       cacheMetrics.record(true, CacheMetrics.SCHEMA_PARTITION_CACHE_NAME);
@@ -802,20 +808,107 @@ public class PartitionCache {
         cacheMetrics.record(false, CacheMetrics.DATA_PARTITION_CACHE_NAME);
         return null;
       }
+
+      final Set<TConsensusGroupId> allConsensusGroupIds = new HashSet<>();
+      final Map<TConsensusGroupId, List<TimeSlotRegionInfo>> consensusGroupToTimeSlotMap =
+          new HashMap<>();
+
       Map<String, Map<TSeriesPartitionSlot, Map<TTimePartitionSlot, List<TRegionReplicaSet>>>>
           dataPartitionMap = new HashMap<>();
-      // check cache for each database
+
       for (Map.Entry<String, List<DataPartitionQueryParam>> entry :
           databaseToQueryParamsMap.entrySet()) {
-        if (null == entry.getValue()
-            || entry.getValue().isEmpty()
-            || !getDatabaseDataPartition(dataPartitionMap, entry.getKey(), entry.getValue())) {
+        String databaseName = entry.getKey();
+        List<DataPartitionQueryParam> params = entry.getValue();
+
+        if (null == params || params.isEmpty()) {
           cacheMetrics.record(false, CacheMetrics.DATA_PARTITION_CACHE_NAME);
           return null;
         }
+
+        DataPartitionTable dataPartitionTable = dataPartitionCache.getIfPresent(databaseName);
+        if (null == dataPartitionTable) {
+          logger.debug(
+              "[{} Cache] miss when search database {}",
+              CacheMetrics.DATA_PARTITION_CACHE_NAME,
+              databaseName);
+          cacheMetrics.record(false, CacheMetrics.DATA_PARTITION_CACHE_NAME);
+          return null;
+        }
+
+        Map<TSeriesPartitionSlot, SeriesPartitionTable> cachedDatabasePartitionMap =
+            dataPartitionTable.getDataPartitionMap();
+        dataPartitionMap.put(databaseName, new HashMap<>());
+
+        for (DataPartitionQueryParam param : params) {
+          TSeriesPartitionSlot seriesPartitionSlot;
+          if (null != param.getDeviceID()) {
+            seriesPartitionSlot = partitionExecutor.getSeriesPartitionSlot(param.getDeviceID());
+          } else {
+            return null;
+          }
+
+          SeriesPartitionTable cachedSeriesPartitionTable =
+              cachedDatabasePartitionMap.get(seriesPartitionSlot);
+          if (null == cachedSeriesPartitionTable) {
+            if (logger.isDebugEnabled()) {
+              logger.debug(
+                  "[{} Cache] miss when search device {}",
+                  CacheMetrics.DATA_PARTITION_CACHE_NAME,
+                  param.getDeviceID());
+            }
+            cacheMetrics.record(false, CacheMetrics.DATA_PARTITION_CACHE_NAME);
+            return null;
+          }
+
+          Map<TTimePartitionSlot, List<TConsensusGroupId>> cachedTimePartitionSlot =
+              cachedSeriesPartitionTable.getSeriesPartitionMap();
+          if (param.getTimePartitionSlotList().isEmpty()) {
+            return null;
+          }
+
+          for (TTimePartitionSlot timePartitionSlot : param.getTimePartitionSlotList()) {
+            List<TConsensusGroupId> cacheConsensusGroupIds =
+                cachedTimePartitionSlot.get(timePartitionSlot);
+            if (null == cacheConsensusGroupIds
+                || cacheConsensusGroupIds.isEmpty()
+                || null == timePartitionSlot) {
+              logger.debug(
+                  "[{} Cache] miss when search time partition {}",
+                  CacheMetrics.DATA_PARTITION_CACHE_NAME,
+                  timePartitionSlot);
+              cacheMetrics.record(false, CacheMetrics.DATA_PARTITION_CACHE_NAME);
+              return null;
+            }
+
+            for (TConsensusGroupId groupId : cacheConsensusGroupIds) {
+              allConsensusGroupIds.add(groupId);
+              consensusGroupToTimeSlotMap
+                  .computeIfAbsent(groupId, k -> new ArrayList<>())
+                  .add(
+                      new TimeSlotRegionInfo(databaseName, seriesPartitionSlot, timePartitionSlot));
+            }
+          }
+        }
       }
+
+      final List<TRegionReplicaSet> allRegionReplicaSets =
+          getRegionReplicaSet(new ArrayList<>(allConsensusGroupIds));
+
+      for (int i = 0; i < allRegionReplicaSets.size(); i++) {
+        TConsensusGroupId groupId = new ArrayList<>(allConsensusGroupIds).get(i);
+        TRegionReplicaSet replicaSet = allRegionReplicaSets.get(i);
+
+        for (TimeSlotRegionInfo info : consensusGroupToTimeSlotMap.get(groupId)) {
+          dataPartitionMap
+              .get(info.databaseName)
+              .computeIfAbsent(info.seriesPartitionSlot, k -> new HashMap<>())
+              .computeIfAbsent(info.timePartitionSlot, k -> new ArrayList<>())
+              .add(replicaSet);
+        }
+      }
+
       logger.debug("[{} Cache] hit", CacheMetrics.DATA_PARTITION_CACHE_NAME);
-      // cache hit
       cacheMetrics.record(true, CacheMetrics.DATA_PARTITION_CACHE_NAME);
       return new DataPartition(dataPartitionMap, seriesSlotExecutorName, seriesPartitionSlotNum);
     } finally {
@@ -823,116 +916,19 @@ public class PartitionCache {
     }
   }
 
-  /**
-   * get dataPartition from database
-   *
-   * @param dataPartitionMap result
-   * @param databaseName database that need to get
-   * @param dataPartitionQueryParams specific query params of data partition
-   * @return whether hit
-   */
-  private boolean getDatabaseDataPartition(
-      Map<String, Map<TSeriesPartitionSlot, Map<TTimePartitionSlot, List<TRegionReplicaSet>>>>
-          dataPartitionMap,
-      String databaseName,
-      List<DataPartitionQueryParam> dataPartitionQueryParams) {
-    DataPartitionTable dataPartitionTable = dataPartitionCache.getIfPresent(databaseName);
-    if (null == dataPartitionTable) {
-      logger.debug(
-          "[{} Cache] miss when search database {}",
-          CacheMetrics.DATA_PARTITION_CACHE_NAME,
-          databaseName);
-      return false;
-    }
-    Map<TSeriesPartitionSlot, SeriesPartitionTable> cachedDatabasePartitionMap =
-        dataPartitionTable.getDataPartitionMap();
-    Map<TSeriesPartitionSlot, Map<TTimePartitionSlot, List<TRegionReplicaSet>>>
-        seriesSlotToTimePartitionMap =
-            dataPartitionMap.computeIfAbsent(databaseName, k -> new HashMap<>());
-    // check cache for each device
-    for (DataPartitionQueryParam dataPartitionQueryParam : dataPartitionQueryParams) {
-      if (!getDeviceDataPartition(
-          seriesSlotToTimePartitionMap, dataPartitionQueryParam, cachedDatabasePartitionMap)) {
-        return false;
-      }
-    }
-    return true;
-  }
+  private static class TimeSlotRegionInfo {
+    final String databaseName;
+    final TSeriesPartitionSlot seriesPartitionSlot;
+    final TTimePartitionSlot timePartitionSlot;
 
-  /**
-   * get dataPartition from device
-   *
-   * @param seriesSlotToTimePartitionMap result
-   * @param dataPartitionQueryParam specific query param of data partition
-   * @param cachedDatabasePartitionMap all cached data partition map of related database
-   * @return whether hit
-   */
-  private boolean getDeviceDataPartition(
-      Map<TSeriesPartitionSlot, Map<TTimePartitionSlot, List<TRegionReplicaSet>>>
-          seriesSlotToTimePartitionMap,
-      DataPartitionQueryParam dataPartitionQueryParam,
-      Map<TSeriesPartitionSlot, SeriesPartitionTable> cachedDatabasePartitionMap) {
-    TSeriesPartitionSlot seriesPartitionSlot;
-    if (null != dataPartitionQueryParam.getDeviceID()) {
-      seriesPartitionSlot =
-          partitionExecutor.getSeriesPartitionSlot(dataPartitionQueryParam.getDeviceID());
-    } else {
-      return false;
+    TimeSlotRegionInfo(
+        String databaseName,
+        TSeriesPartitionSlot seriesPartitionSlot,
+        TTimePartitionSlot timePartitionSlot) {
+      this.databaseName = databaseName;
+      this.seriesPartitionSlot = seriesPartitionSlot;
+      this.timePartitionSlot = timePartitionSlot;
     }
-    SeriesPartitionTable cachedSeriesPartitionTable =
-        cachedDatabasePartitionMap.get(seriesPartitionSlot);
-    if (null == cachedSeriesPartitionTable) {
-      if (logger.isDebugEnabled()) {
-        logger.debug(
-            "[{} Cache] miss when search device {}",
-            CacheMetrics.DATA_PARTITION_CACHE_NAME,
-            dataPartitionQueryParam.getDeviceID());
-      }
-      return false;
-    }
-    Map<TTimePartitionSlot, List<TConsensusGroupId>> cachedTimePartitionSlot =
-        cachedSeriesPartitionTable.getSeriesPartitionMap();
-    Map<TTimePartitionSlot, List<TRegionReplicaSet>> timePartitionSlotListMap =
-        seriesSlotToTimePartitionMap.computeIfAbsent(seriesPartitionSlot, k -> new HashMap<>());
-    // Notice: when query all time partition, then miss
-    if (dataPartitionQueryParam.getTimePartitionSlotList().isEmpty()) {
-      return false;
-    }
-    // check cache for each time partition
-    for (TTimePartitionSlot timePartitionSlot :
-        dataPartitionQueryParam.getTimePartitionSlotList()) {
-      if (!getTimeSlotDataPartition(
-          timePartitionSlotListMap, timePartitionSlot, cachedTimePartitionSlot)) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  /**
-   * get dataPartition from time slot
-   *
-   * @param timePartitionSlotListMap result
-   * @param timePartitionSlot the specific time partition slot of data partition
-   * @param cachedTimePartitionSlot all cached time slot map of related device
-   * @return whether hit
-   */
-  private boolean getTimeSlotDataPartition(
-      Map<TTimePartitionSlot, List<TRegionReplicaSet>> timePartitionSlotListMap,
-      TTimePartitionSlot timePartitionSlot,
-      Map<TTimePartitionSlot, List<TConsensusGroupId>> cachedTimePartitionSlot) {
-    List<TConsensusGroupId> cacheConsensusGroupIds = cachedTimePartitionSlot.get(timePartitionSlot);
-    if (null == cacheConsensusGroupIds
-        || cacheConsensusGroupIds.isEmpty()
-        || null == timePartitionSlot) {
-      logger.debug(
-          "[{} Cache] miss when search time partition {}",
-          CacheMetrics.DATA_PARTITION_CACHE_NAME,
-          timePartitionSlot);
-      return false;
-    }
-    timePartitionSlotListMap.put(timePartitionSlot, getRegionReplicaSet(cacheConsensusGroupIds));
-    return true;
   }
 
   /**