# Motivation

We found that on physical machines with good configuration, if there are a large number of regions and devices, and each write operation writes only one row of data per device, the PartitionCache's regionLock becomes a bottleneck. Although all operations are using read locks, on a machine with 96 cores and 192 threads, high-concurrency read operations involving CAS (Compare-And-Swap) on a read lock still turn into a bottleneck.

To address this, we batched the logic, reducing the number of times each thread needs to acquire a read lock from region num to just once overall. This eliminated the read lock bottleneck.

# Configuration
## Machine：
* 96C 192 Thread 768GB Memory
## IoTDB Server：1C1D
* data_region_group_extension_policy=CUSTOM
* default_data_region_group_num_per_database=192
## Benchmark：
* LOOP=100
* DEVICE_NUMBER=4500000
* SENSOR_NUMBER=2
* DATA_CLIENT_NUMBER=150
* POINT_STEP=1000
* OP_MIN_INTERVAL=-1
* INSERT_DATATYPE_PROPORTION=0:1:0:1:0:0:0:0:0:0
* BATCH_SIZE_PER_WRITE=1
* DEVICE_NUM_PER_WRITE=30000
* IS_OUT_OF_ORDER=false

# Test result
## Before
```
Create schema cost 3.48 second
Test elapsed time (not include schema creation): 360.23 second
----------------------------------------------------------Result Matrix----------------------------------------------------------
Operation                okOperation              okPoint                  failOperation            failPoint                throughput(point/s)
INGESTION                15000                    900000000                0                        0                        2498376.56
PRECISE_POINT            0                        0                        0                        0                        0.00
TIME_RANGE               0                        0                        0                        0                        0.00
VALUE_RANGE              0                        0                        0                        0                        0.00
AGG_RANGE                0                        0                        0                        0                        0.00
AGG_VALUE                0                        0                        0                        0                        0.00
AGG_RANGE_VALUE          0                        0                        0                        0                        0.00
GROUP_BY                 0                        0                        0                        0                        0.00
LATEST_POINT             0                        0                        0                        0                        0.00
RANGE_QUERY_DESC         0                        0                        0                        0                        0.00
VALUE_RANGE_QUERY_DESC   0                        0                        0                        0                        0.00
GROUP_BY_DESC            0                        0                        0                        0                        0.00
---------------------------------------------------------------------------------------------------------------------------------

--------------------------------------------------------------------------Latency (ms) Matrix--------------------------------------------------------------------------
Operation                AVG         MIN         P10         P25         MEDIAN      P75         P90         P95         P99         P999        MAX         SLOWEST_THREAD
INGESTION                3419.17     307.01      2434.31     2908.19     3326.67     3678.20     3991.60     4208.49     13318.90    20724.98    21227.74    351047.16
PRECISE_POINT            0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
TIME_RANGE               0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
VALUE_RANGE              0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
AGG_RANGE                0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
AGG_VALUE                0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
AGG_RANGE_VALUE          0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
GROUP_BY                 0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
LATEST_POINT             0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
RANGE_QUERY_DESC         0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
VALUE_RANGE_QUERY_DESC   0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
GROUP_BY_DESC            0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

<img width="1695" alt="image" src="https://github.com/user-attachments/assets/d80d2138-b72e-4eea-b52c-e9751e131f86" />

## After
```
Create schema cost 2.69 second
Test elapsed time (not include schema creation): 165.53 second
----------------------------------------------------------Result Matrix----------------------------------------------------------
Operation                okOperation              okPoint                  failOperation            failPoint                throughput(point/s)
INGESTION                15000                    900000000                0                        0                        5437068.69
PRECISE_POINT            0                        0                        0                        0                        0.00
TIME_RANGE               0                        0                        0                        0                        0.00
VALUE_RANGE              0                        0                        0                        0                        0.00
AGG_RANGE                0                        0                        0                        0                        0.00
AGG_VALUE                0                        0                        0                        0                        0.00
AGG_RANGE_VALUE          0                        0                        0                        0                        0.00
GROUP_BY                 0                        0                        0                        0                        0.00
LATEST_POINT             0                        0                        0                        0                        0.00
RANGE_QUERY_DESC         0                        0                        0                        0                        0.00
VALUE_RANGE_QUERY_DESC   0                        0                        0                        0                        0.00
GROUP_BY_DESC            0                        0                        0                        0                        0.00
---------------------------------------------------------------------------------------------------------------------------------

--------------------------------------------------------------------------Latency (ms) Matrix--------------------------------------------------------------------------
Operation                AVG         MIN         P10         P25         MEDIAN      P75         P90         P95         P99         P999        MAX         SLOWEST_THREAD
INGESTION                1424.66     294.88      848.60      1029.60     1214.31     1397.61     1566.03     1684.30     10181.36    18736.46    19026.30    151111.42
PRECISE_POINT            0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
TIME_RANGE               0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
VALUE_RANGE              0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
AGG_RANGE                0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
AGG_VALUE                0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
AGG_RANGE_VALUE          0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
GROUP_BY                 0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
LATEST_POINT             0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
RANGE_QUERY_DESC         0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
VALUE_RANGE_QUERY_DESC   0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
GROUP_BY_DESC            0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00        0.00
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

<img width="1635" alt="image" src="https://github.com/user-attachments/assets/3611db4a-ee32-4d65-81c5-724fbbb85836" />


# Conclusion

With this batch optimization, we removed the regionLock bottleneck, improving write performance by 112% in this case.

BTW, We also checked all locks on PartitionCache to ensure that no other locks would have similar workload-dependent bindings and become bottlenecked.

# Future Optimization

We may optimize write performance in this case by futher optimizing validateDeviceSchemaInCache